# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Civil Protection Department and Ministry of Health official COVID19 data in Italy. More info <a target='_blank' href='https://github.com/alessiodl/COVID19Feed2Json'>here</a>",
      "version": "1.0.0",
      "title": "COVID19-API",
      "contact": {
        "email": "alessio.dl@gmail.com"
      },
      "license": {
        "name": "CC BY 4.0",
        "url": "https://creativecommons.org/licenses/by/4.0/deed.it"
      }
    },
    "schemes":["http","https"],
    "produces":["application/json"],
    "paths":{
        "/info":{
            "get":{
                "summary":"General Informations about the API",
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/andamento":{
            "get":{
                "summary":"Daily updates about the disease spread",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": false,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-03-08 18:00:00"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/regioni":{
            "get":{
                "summary":"Daily updates about infection distribution at regional level in GeoJSON Point format",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": false,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-03-08"
                },{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "3"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/regioni/map":{
            "get":{
                "summary":"Daily updates about infection distribution at regional level in GeoJSON Polygon format",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": true,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-03-08"
                },{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "3"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/province":{
            "get":{
                "summary":"Daily updates about infection distribution at provinces level in GeoJSON Point format",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": false,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-03-08"
                },{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "3"
                },{
                    "name": "cod_prov",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the province",
                    "type": "string",
                    "example": "15"
                },{
                    "name": "sigla_prov",
                    "in": "query",
                    "required": false,
                    "description": "province short name",
                    "type": "string",
                    "example": "MI"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/province/map":{
            "get":{
                "summary":"Daily updates about infection distribution at province level in GeoJSON Polygon format",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": true,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-03-08"
                },{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "3"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/comuni":{
            "get":{
                "summary":"Daily updates about infection distribution at municipalities level in JSON format",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": false,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-03-18"
                },{
                    "name": "sigla_prov",
                    "in": "query",
                    "required": false,
                    "description": "Province short name, also as comma separated list",
                    "type": "string",
                    "example": "TE"
                },{
                    "name": "cod_istat",
                    "in": "query",
                    "required": false,
                    "description": "Municipality ISTAT code",
                    "type": "string",
                    "example": "67025"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/comuni/map":{
            "get":{
                "summary":"Daily updates about infection distribution at municipalities level in GeoJSON Polygon format",
                "parameters": [{
                    "name": "data",
                    "in": "query",
                    "required": true,
                    "description": "Reference date",
                    "type": "string",
                    "example": "2020-04-21"
                },{
                    "name": "sigla_prov",
                    "in": "query",
                    "required": false,
                    "description": "Province short name",
                    "type": "string",
                    "example": "TE"
                },{
                    "name": "cod_istat",
                    "in": "query",
                    "required": false,
                    "description": "Municipality ISTAT code",
                    "type": "string",
                    "example": "67025"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/asl/totals":{
            "get":{
                "summary":"Summary data about IZSAM analysis by ASL",
                "parameters":[],
                "responses":{
                    "200":{
                        "description":"OK"
                    }
                }
            }
        },
        "/asl/daily":{
            "get":{
                "summary":"Daily updates about IZSAM analysis by ASL",
                "parameters":[],
                "responses":{
                    "200":{
                        "description":"OK"
                    }
                }
            }
        },
        "/ospedali/covid19":{
            "get":{
                "summary":"COVID-19 hospitals GeoJSON points",
                "parameters": [{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "14"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/ospedali/covid19/post":{
            "get":{
                "summary":"Post acute COVID-19 hospitals GeoJSON points",
                "parameters": [{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "14"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        },
        "/ospedali/covid19/altri":{
            "get":{
                "summary":"Not COVID-19 specifical structures GeoJSON points",
                "parameters": [{
                    "name": "cod_reg",
                    "in": "query",
                    "required": false,
                    "description": "ISTAT code of the region",
                    "type": "string",
                    "example": "14"
                }],
                "responses": {
                    "200": {
                      "description": "OK"
                    }
                }
            }
        }
    }
    
}